add basic test and refactor more
