+++ app/packages/state/src/hooks/useSchemaSettings.test.ts
@@ -1,5 +1,5 @@
 import { afterEach, describe, expect, it, vi } from "vitest";
-import { disabledField } from "./useSchemaSettings";
+import { disabledField } from "./useSchemaSettings.utils";
 import { OBJECT_ID_FIELD } from "@fiftyone/utilities";
 
 const FIELDS = {
@@ -17,6 +17,9 @@ const FIELDS = {
   },
 };
 
+const GROUP_DATASET = "group";
+const NOT_GROUP_DATASET = "";
+
 describe("Disabled schema fields", () => {
   afterEach(() => {
     vi.restoreAllMocks();
@@ -26,7 +29,7 @@ describe("Disabled schema fields", () => {
     const path = "id";
     const field_1 = FIELDS.ID_FIELD;
     const schema = { id: field_1 };
-    const isGroupDataset = "group";
-    expect(disabledField(path, schema, isGroupDataset)).toBe(true);
+
+    expect(disabledField(path, schema, NOT_GROUP_DATASET)).toBe(true);
   });
 });
